# resttemplate-demo/person-server/src/main/java/com/example/personserver/api/endpoint/SwaggerConstants.java
package com.example.personserver.api.endpoint;

import com.example.personserver.api.dto.PersonDto;
import org.springframework.data.domain.Page;

public class SwaggerConstants {

    interface PersonPage extends Page<PersonDto> {}
}